Rename consumer.py and provider.py to be less confusing
